rename sigma to metasag
