+++ libmetasag.py
